Refactor logic into common code.
